+++ src/api/getStatsOfObjectsToUpload.ts
@@ -1,0 +1,13 @@
+import {ObjectStatsMap, FileObjectStats} from './types';
+
+export function getStatsOfObjectsToUpload(
+  keysOfObjectsToUpload: string[],
+  statsOfObjectsOnDisk: ObjectStatsMap<FileObjectStats>
+): ObjectStatsMap<FileObjectStats> {
+  return keysOfObjectsToUpload.reduce((accum, key) => {
+    return {
+      ...accum,
+      [key]: statsOfObjectsOnDisk[key]
+    };
+  }, {});
+}